requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.6
Django==4.2
django-countries==7.5.1
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-simple-history==3.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
filelock==3.11.0
gunicorn==20.1.0
identify==2.5.22
mysqlclient==2.1.1
nodeenv==1.7.0
Pillow==9.4.0
platformdirs==3.2.0
pre-commit==3.2.2
psycopg2==2.9.6
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
typing-extensions==4.5.0
virtualenv==20.21.0
